Fixed bug where selection set does not work.
When a user does not have a WorldEdit LocalSession yet and a sk_posX()
call is made, it is possible that WE creates a new LocalSession without
storing it, resulting in a 'floating' session that's not used by
anything. In other words, sk_posX() will not set the selection. This bug
rarely occurs, but if it does, it is after a server restart on the first
join of a player. The other commits in the PR do not affect this bug.